:recycle: refactor: validator 분리
