+++ src/validator/productValidator.ts
@@ -1,0 +1,22 @@
+import { Product } from '../domain/Product';
+import { ERROR_MESSAGE, CONFIGURATION } from '../constants';
+
+const productValidator = {
+  isDuplicated(name: string, products: Product[]) {
+    return products.some((product) => product.name === name);
+  },
+
+  isIncorrectUnit(price: number) {
+    return price % CONFIGURATION.PRICE.UNIT !== 0;
+  },
+};
+
+export const validateProduct = (product: Product, products: Product[]) => {
+  if (productValidator.isDuplicated(product.name, products)) {
+    throw new Error(ERROR_MESSAGE.DUPLICATED_PRODUCT);
+  }
+
+  if (productValidator.isIncorrectUnit(product.price)) {
+    throw new Error(ERROR_MESSAGE.INCORRECT_UNIT_PRODUCT_PRICE);
+  }
+};
